make module for modal
