+++ scripts/main.js
@@ -1,9 +1,8 @@
 import myLibrary from "./library.js";
-import Book from "./book.js";
-import { renderBooks,renderBooksInfo } from "./helper.js";
+import modal from "./modal.js";
+import { renderBooks, renderBooksInfo } from "./helper.js";
 
 /* library */
-
 const library = document.createElement("div");
 document.getElementsByTagName("body")[0].appendChild(library);
 library.classList.add("flex", "flex-row", "min-h-full");
@@ -27,12 +26,14 @@ bookTable.classList.add(
 /* aside component */
 
 const info = document.createElement("aside");
+info.id = "info";
 info.classList.add("bg-red-400", "border-l-2", "w-1/4", "pt-20");
 info.innerHTML = "<h3 class='text-center'>Information</h3>";
 info.innerHTML += "<hr>";
-info.innerHTML += '<div class="flex justify-between px-4 py-2"><p>Books</p><p id="info-total"></p></div>';
-info.innerHTML += '<div class="flex justify-between px-4 py-2"><p>Completed Books</p><p id="info-finished"></p></div>'; 
-
+info.innerHTML +=
+  '<div class="flex justify-between px-4 py-2"><p>Books</p><p id="info-total"></p></div>';
+info.innerHTML +=
+  '<div class="flex justify-between px-4 py-2"><p>Completed Books</p><p id="info-finished"></p></div>';
 
 // -----------------
 
@@ -43,21 +44,6 @@ library.appendChild(info);
 /* header */
 let addBookModal = document.getElementById("add-book-modal");
 
-/* modal elements */
-
-let modal = document.getElementById("modal");
-let modalCancelButton = modal.getElementsByTagName("button")[1];
-let modalAddButton = modal.getElementsByTagName("button")[0];
-
-let title = document.getElementById("title");
-let author = document.getElementById("author");
-let genre = document.getElementById("genre");
-let pages = document.getElementById("pages");
-let read_pages = document.getElementById("read_pages");
-let isRead = document.getElementById("isRead");
-
-// --------------------
-
 /* Render books */
 
 renderBooks(myLibrary, bookTable);
@@ -67,26 +53,4 @@ renderBooksInfo(myLibrary, info);
 
 addBookModal.addEventListener("click", function (e) {
   modal.classList.remove("hidden");
-});
-
-modalAddButton.addEventListener("click", function (e) {
-  e.preventDefault;
-  let book = new Book(
-    title.value,
-    author.value,
-    genre.value,
-    pages.value,
-    read_pages.value,
-    cover.value,
-    isRead.value
-  );
-  myLibrary.addBookToLibrary(book);
-  bookTable.innerHTML = "";
-  renderBooks(myLibrary, bookTable);
-  renderBooksInfo(myLibrary, info);
-  modal.classList.add("hidden");
-});
-
-modalCancelButton.addEventListener("click", () => {
-  modal.classList.add("hidden");
 });
